requirements.txt
yadisk[sync_defaults]
pillow
requests